customer-wise daily report

+++ app/src/main/java/com/oxvsys/moneylender/FragmentDashboard.java
@@ -96,7 +96,7 @@ public class FragmentDashboard extends Fragment {
                                 Account account_temp = new Account();
                                 account_temp.setNo(account.getKey());
                                 accountAmountCollectList.add(new AccountAmountCollect(account_temp,
-                                        Long.parseLong(account.getValue().toString())));
+                                        Integer.parseInt(account.getValue().toString())));
 
 
 //                                total_daily_amount += Long.parseLong(account.getValue().toString());
